Update imports for proxy-codegen and backend-function-client renames
